requirements.txt
streamlit
google-generativeai
python-dotenv
requests
Pillow
openai
PyPDF2
python-docx